+++ src/Mvx.ApiClient.Net/Clients/NetworkClient.cs
@@ -12,23 +12,23 @@ internal sealed class NetworkClient : INetworkClient
         _requestExecutor = new ApiRequestExecutor(httpClient);
     }
 
-    public async Task<StatsDto> GetStatsAsync(DataSelection? dataSelection = null, CancellationToken cancellationToken = default)
+    public async Task<StatsDto> GetStatsAsync(CancellationToken cancellationToken = default)
     {
-        return await _requestExecutor.GetAsync<StatsDto>(EndpointPaths.NetworkStats, new QueryOptions { Data = dataSelection }, cancellationToken);
+        return await _requestExecutor.GetAsync<StatsDto>(EndpointPaths.NetworkStats, null, cancellationToken);
     }
 
-    public async Task<EconomicsDto> GetEconomicsAsync(DataSelection? dataSelection = null, CancellationToken cancellationToken = default)
+    public async Task<EconomicsDto> GetEconomicsAsync(CancellationToken cancellationToken = default)
     {
-        return await _requestExecutor.GetAsync<EconomicsDto>(EndpointPaths.NetworkEconomics, new QueryOptions { Data = dataSelection }, cancellationToken);
+        return await _requestExecutor.GetAsync<EconomicsDto>(EndpointPaths.NetworkEconomics, null, cancellationToken);
     }
 
-    public async Task<NetworkConstantsDto> GetConstantsAsync(DataSelection? dataSelection = null, CancellationToken cancellationToken = default)
+    public async Task<NetworkConstantsDto> GetConstantsAsync(CancellationToken cancellationToken = default)
     {
-        return await _requestExecutor.GetAsync<NetworkConstantsDto>(EndpointPaths.NetworkConstants, new QueryOptions { Data = dataSelection }, cancellationToken);
+        return await _requestExecutor.GetAsync<NetworkConstantsDto>(EndpointPaths.NetworkConstants, null, cancellationToken);
     }
 
-    public async Task<AboutDto> GetAboutAsync(DataSelection? dataSelection = null, CancellationToken cancellationToken = default)
+    public async Task<AboutDto> GetAboutAsync(CancellationToken cancellationToken = default)
     {
-        return await _requestExecutor.GetAsync<AboutDto>(EndpointPaths.NetworkAbout, new QueryOptions { Data = dataSelection }, cancellationToken);
+        return await _requestExecutor.GetAsync<AboutDto>(EndpointPaths.NetworkAbout, null, cancellationToken);
     }
 }