+++ src/Mvx.ApiClient.Net/Constants.cs
@@ -2,7 +2,7 @@
 
 internal static class Constants
 {
-    internal const string BaseAddressMainnetApi = "https://api.multiversx.com";
-    internal const string BaseAddressTestnetApi = "https://testnet-api.multiversx.com";
-    internal const string BaseAddressDevnetApi = "https://devnet-api.multiversx.com";
+    internal const string BaseAddressMainnetApi = "https://api.multiversx.com/";
+    internal const string BaseAddressTestnetApi = "https://testnet-api.multiversx.com/";
+    internal const string BaseAddressDevnetApi = "https://devnet-api.multiversx.com/";
 }